added pagination components to OtherImages component
