+++ src/components/UserDashboard/OtherUserDashboard/OtherUserProfile.jsx
@@ -1,28 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import userImage from '../../../assets/images/user.svg';
-import { api } from '../../../api';
-import { isAuthenticated, setAuthToken } from '../../Landing/Auth';
 
-const OtherUserProfile = ({ userId }) => {
-  const [user, setUser] = useState(null);
-  const { accessToken } = isAuthenticated();
-  const [followee, setFollowee] = useState(0);
-  const [follower, setFollower] = useState(0);
-  const [posts, setPosts] = useState(0);
-
-  useEffect(() => {
-    setAuthToken(accessToken);
-    const fetchUser = async () => {
-      const response = await api.get(`/api/user/${userId}`);
-      setUser(response.data[0]);
-      setPosts(response.data[0].Photos[0].posts);
-      setFollowee(response.data[1].followee.length);
-      setFollower(response.data[2].follower.length);
-    };
-    fetchUser();
-  }, [accessToken, userId]);
-
+const OtherUserProfile = ({ user, followee, follower, posts }) => {
   return (
     <ProfileWrapper>
       {user ? (
